mca_insights_engine/requirements.txt
pandas==2.2.2
numpy==1.26.4
sqlalchemy==2.0.32
python-dotenv==1.0.1
tqdm==4.66.4
requests==2.32.3
beautifulsoup4==4.12.3
streamlit==1.37.1
Flask==3.0.3
